updated Timer to be safer to run by using SafeRead methods vs try/catch blocks

+++ src/utils/Timer.ts
@@ -1,14 +1,19 @@
 import { FileSystem } from "FileSystem/FileSystem"
+import { SafeReadFromFile } from "./UtilFuncs"
 
 export class Timer {
     private static all_timers: Timer[] = []
 
     private timer_id: string
     private path: string[]
+    private timer_end_key: string
+    private timer_current_key: string
 
     constructor(timer_id: string) {
         this.timer_id = timer_id
         this.path = ['timers', this.timer_id]
+        this.timer_end_key = "timer_limit"
+        this.timer_current_key = "timer_start"
 
         if (!Timer.all_timers.includes(this)) {
             Timer.all_timers.push(this)
@@ -23,41 +28,36 @@ export class Timer {
     }
 
     private TickTimer() {
-        try {
-            const FILE = FileSystem.GetFileSystem().GetExistingFile(this.path)
-            const CUR_TIME = Number(FILE?.ReadFromFile('timer_start'))
-            const MAX_TIME = Number(FILE?.ReadFromFile('timer_limit'))
-            if (CUR_TIME >= MAX_TIME) {
-                FileSystem.GetFileSystem().DeleteFile(this.path)
-                return
-            }
-            FILE?.WriteToFile('timer_start', Number(CUR_TIME) + 1)
-        }
-        catch {
+        const FILE = FileSystem.GetFileSystem().GetExistingFile(this.path)
+        if (FILE == null) { return }
+
+        const CUR_TIME = Number(SafeReadFromFile(FILE, this.timer_current_key))
+        const MAX_TIME = Number(SafeReadFromFile(FILE, this.timer_end_key))
+
+        if (CUR_TIME >= MAX_TIME) {
             FileSystem.GetFileSystem().DeleteFile(this.path)
+            return
         }
+
+        FILE.WriteToFile(this.timer_current_key, CUR_TIME + 1)
     }
 
     public StartTimer(tick_time: number) {
         const FILE_SYSTEM = FileSystem.GetFileSystem()
         if (!FILE_SYSTEM.DoesFileExist(this.path)) {
             const FILE = FileSystem.GetFileSystem().GetFile(this.path)
-            FILE.WriteToFile('timer_limit', tick_time)
-            FILE.WriteToFile('timer_start', 0)
+            FILE.WriteToFile(this.timer_end_key, tick_time)
+            FILE.WriteToFile(this.timer_current_key, 0)
         }
     }
 
     public IsTimerDone() {
         const FILE = FileSystem.GetFileSystem().GetExistingFile(this.path)
+        if (FILE == null) { return false }
 
-        try {
-            const TIME_LIMIT = FILE!.ReadFromFile('timer_limit')
-            const CUR_TIME = FILE!.ReadFromFile('timer_start')
-            const DONE = Number(CUR_TIME) >= Number(TIME_LIMIT)
-            return DONE
-        }
-        catch {
-            return false
-        }
+        const TIME_LIMIT = Number(SafeReadFromFile(FILE, this.timer_end_key))
+        const CUR_TIME = Number(SafeReadFromFile(FILE, this.timer_current_key))
+
+        return CUR_TIME >= TIME_LIMIT
     }
 }
